try 1

+++ main_translate_rnn.py
@@ -92,7 +92,7 @@ def main():
             optimizer.zero_grad()
 
             # pdb.set_trace()
-            src, trg = one_hotty(batch).permute(1, 0, 2, 3).to(torch.half)
+            src, trg = one_hotty(batch).permute(1, 0, 2, 3).to(torch.long)
 
             x = model(src)
